finsihed refactor
