+++ Bubble/SceneManager.swift
@@ -22,83 +22,6 @@ class SceneManager: NSObject, ARSCNViewDelegate {
         }
     }
     
-//    func updateText(text: String, in sceneView: ARSCNView) {
-//
-//        let (direction, _) = getUserVector(in: sceneView)
-//        
-//        var positionX: Float
-//        var positionZ: Float
-//        
-//        let distance: Float = 2.0
-//
-//        // If there's a previous direction, compare it with the current direction
-//        if let originalDirection = originalOrientation {
-//            // Calculate the angle difference between the two directions
-//            let angleDifference = angleBetweenVectors(vectorA: originalDirection, vectorB: direction)
-//
-//            let angleDifferenceDegrees = angleDifference * 180 / .pi
-//
-//            // Determine the direction of rotation
-//            let crossProduct = crossProduct(vectorA: originalDirection, vectorB: direction)
-//            
-//            let clockwise = crossProduct.z > 0
-//            if clockwise {
-//                print("\(angleDifferenceDegrees) clockwise (right)")
-//            } else {
-//                print("\(angleDifferenceDegrees) counter clockwise (left)")
-//            }
-//            
-//            
-//            
-//            if clockwise {
-//                positionX = -1 * sin(-1 * angleDifference)
-//                positionZ = -1 * cos(-1 * angleDifference)
-//            } else {
-//                positionX = -1 * sin(angleDifference)
-//                positionZ = -1 * cos(angleDifference)
-//            }
-//            
-//            
-//            
-//            print(positionX, positionZ)
-//            
-//            // create text geometry
-//            let textGeometry = SCNText(string: insertNewlines(string: text, every: 20), extrusionDepth: 1.0)
-//            textGeometry.firstMaterial?.diffuse.contents = UIColor.red
-//            
-//            textNode.removeFromParentNode()
-//
-//            textNode = SCNNode(geometry: textGeometry)
-//            textNode.position = SCNVector3(positionX * distance, 0, positionZ * distance)
-//            textNode.scale = SCNVector3(0.01, 0.01, 0.01)
-//            
-//            
-//            // Add a billboard constraint to make the text node always face the camera
-//            let billboardConstraint = SCNBillboardConstraint()
-//            billboardConstraint.freeAxes = SCNBillboardAxis.Y
-//            textNode.constraints = [billboardConstraint]
-//        }
-//
-//        
-//
-//        // Create a box that will act as a background for the text
-//        let (min, max) = textNode.boundingBox
-//        
-//        let boxWidth = CGFloat(max.x - min.x) + 5
-//        let boxHeight = CGFloat(max.y - min.y) + 3
-//        let box = SCNBox(width: boxWidth, height: boxHeight, length: CGFloat(max.z - min.z), chamferRadius: 0.0)
-//        box.firstMaterial?.diffuse.contents = UIColor.white
-//
-//        // Create a node for the box and position it behind the text
-//        let boxNode = SCNNode(geometry: box)
-//        boxNode.position = SCNVector3(CGFloat(textNode.position.x) + (boxWidth * 0.45) , CGFloat(textNode.position.y) + 6.3, -0.01) // Adjust the z-position to move the box behind the text
-//
-//        // Add the box node as a child of the text node
-//        textNode.addChildNode(boxNode)
-//
-//        sceneView.scene.rootNode.addChildNode(textNode)
-//    }
-    
     func updateText(text: String, in sceneView: ARSCNView) {
         let (direction, _) = getUserVector(in: sceneView)
         let distance: Float = 2.0
